pnpm workspaces, package updates, removing buffer (#88)
I used pnpm to work out the version incompatibilities. Using `pnpm` (or
npm or yarn) workspaces instead of the `cd frontend` pattern worked
better for me.

I had to upgrade vite, and then Buffer became a pain, so I replaced it
with the `btoa` api. This should work, and the build now only rejects
when it searches for production canister_ids

---------

Co-authored-by: Islam El-Ashi <islam.elashi@dfinity.org>